Create ANTLR4 Parser for proto3
This commit adds a Protobuf 3 parser using Buf's
[ANTLR4 .g4](https://github.com/bufbuild/protobuf.com/tree/main/examples/antlr)
lexer and parser files.

The code contains a PDM script block to generate the Python 3 files. The
generated files are gitignored. They're meant to be built prior to publication
to pypi.

The code also includes a parser and model file, which contain a little helper to
parse .proto text into data classes. This implementation is currently very
basic, but works for simple stuff.