Default cardinality to None
I decided to reflect what the CST contains for the FieldCardinality in fields
and groups. If unset, it now uses None, rather than OPTIONAL. I did this because
Protobuf's rules for unset and implict defaults and stuff are pretty
complicated. I'd rather leave that up to the developers reading the AST.