Support message literals (#41)
I've added full support for message literals in option values. This
involved adding a couple of new AST types:

- Added `MessageLiteral`, `MessageLiteralField`, and `MessageValue`
types.
- Updated `Option` to contain either `ScalaerValue` or `MessageLiteral`

Message literals contain a list of fields. Each field has a name and a
value. The value can be a scalar value (1, 2, "foo", true), or a message
literal (nested). It can also be a list (`[1, 2, <some message
literal>]`).