Add support for `listOfMessagesLiteral` in `messgaeValue` (#46)
The parser didn't support a message literal made up of a list of message
literals. This is valid syntax that's captured in the `.g4` and used by
protobuf definitions such as etcd (see #45). The parser is updated to
handle this style. I've also added a test to the generator to verify
that everything works as expected, which it does.

Fixes #45