Add better support for inline comments (#65)
This pull request...

1. introduces support for inline comments in the AST and code generation
for `.proto` schema parsing and generation.
2. refactors how message literals are represented, allowing them to
contain both fields and comments, and updates the generator logic and
tests accordingly.
3. improves the formatting of reserved values and fixes some minor
issues in the test suite.

I've elected to make (2) a backwards incompatible change (`fields` is
replaced with `elements`). I plan to release this change as a 2.0.0
version.

Fixes #62